add d3 svg outputs container
