+++ components/displays/output_item.jsx
@@ -26,33 +26,33 @@ class OutputItem extends React.Component {
             this.trainingData.push(obj);
         })
 
-        const svg = d3.select("svg");
-        console.log("svg", svg);
+        // const svg = d3.select("svg");
+        // console.log("svg", svg);
 
-        let node = {
-            string: output,
-            x: Math.random() * 800,
-            y: Math.random() * 200,
-        }
+        // let node = {
+        //     string: output,
+        //     x: Math.random() * 800,
+        //     y: Math.random() * 200,
+        // }
 
-        let nodeArray = [];
+        // let nodeArray = [];
 
-        nodeArray.push(node);
+        // nodeArray.push(node);
 
-        svg.append("g")
-            .attr('transform', 'translate(30,30)')
+        // svg.append("g")
+        //     .attr('transform', 'translate(30,30)')
         
-        const circle = svg.selectAll("circle")
-            .data(node)
-        const g = circle.enter().append("g")
-            g.append("circle")
-            .attr('r', 40)
-            .attr('cy', (d) => d.y )
-            .attr('cx', (d) => d.x )
-            .attr('class', 'node')
-            .style('fill', 'white')
-            .style('stroke-width', '5px')
-            .style('stroke', "blue")
+        // const circle = svg.selectAll("circle")
+        //     .data(node)
+        // const g = circle.enter().append("g")
+        //     g.append("circle")
+        //     .attr('r', 40)
+        //     .attr('cy', (d) => d.y )
+        //     .attr('cx', (d) => d.x )
+        //     .attr('class', 'node')
+        //     .style('fill', 'white')
+        //     .style('stroke-width', '5px')
+        //     .style('stroke', "blue")
